requirements.txt
streamlit>=1.30
langchain==0.2.14
langchain-google-genai==1.0.6
python-dotenv
